[NDGL-100] design: 공통되는 컴포넌트 정리 - AlternativePlaceContent, PlaceDetailTabRow, PlaceInfoTab, PlacePhotoTab, PlaceTipCard
