--- feature/travel/src/main/java/com/yapp/ndgl/feature/travel/placedetail/component/PlaceInfoRow.kt
@@ -1,51 +1,0 @@
-package com.yapp.ndgl.feature.travel.placedetail.component
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.RowScope
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material3.Icon
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.unit.dp
-import com.yapp.ndgl.core.ui.R
-import com.yapp.ndgl.core.ui.theme.NDGLTheme
-import com.yapp.ndgl.core.ui.util.noRippleClickable
-
-@Composable
-internal fun PlaceInfoRow(
-    iconRes: Int,
-    onClick: (() -> Unit)? = null,
-    content: @Composable RowScope.() -> Unit,
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .then(if (onClick != null) Modifier.noRippleClickable { onClick() } else Modifier),
-        horizontalArrangement = Arrangement.spacedBy(8.dp),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Icon(
-            imageVector = ImageVector.vectorResource(iconRes),
-            contentDescription = null,
-            tint = NDGLTheme.colors.green500,
-        )
-        content()
-        if (onClick != null) {
-            Icon(
-                modifier = Modifier
-                    .size(24.dp)
-                    .clip(CircleShape),
-                imageVector = ImageVector.vectorResource(R.drawable.ic_24_chevron_right),
-                tint = NDGLTheme.colors.black600,
-                contentDescription = null,
-            )
-        }
-    }
-}
